requirements.txt
python-telegram-bot==21.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
Pillow>=10.0.0
